# @Osama/@Osama/backend/openapi.yaml
openapi: 3.0.0
info:
  title: Kids Safe Digital Library API
  version: 1.0.0
  description: 'API لمكتبة أطفال رقمية آمنة: قصص، فيديوهات، ألعاب. واجهة عامة للقراءة فقط + لوحة إدارة (JWT). لا حسابات مستخدمين عامة، لا تتبع أو تحليلات.'
servers:
  - url: http://localhost:4000/api
    description: Development
  - url: /api
    description: Relative (same host)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: توكن من POST /admin/login
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          description: رسالة خطأ قابلة للقراءة
      description: استجابة الخطأ الموحدة
    ContentPage:
      type: object
      properties:
        pageNumber:
          type: integer
          minimum: 1
        imageUrl:
          type: string
          format: uri
        text:
          type: string
          nullable: true
      description: صفحة من محتوى (قصة)
    Content:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum:
            - story
            - video
            - game
        ageMin:
          type: integer
          minimum: 0
        ageMax:
          type: integer
          minimum: 0
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        contentUrl:
          type: string
          format: uri
          nullable: true
          description: فيديو أو رابط خارجي
        fileUrl:
          type: string
          format: uri
          nullable: true
          description: ملف PDF للألعاب (sourceType=uploaded)
        sourceType:
          type: string
          enum:
            - uploaded
            - youtube
            - external
        isActive:
          type: boolean
          default: true
        orderIndex:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryItem'
        ageGroups:
          type: array
          items:
            $ref: '#/components/schemas/AgeGroupItem'
        pages:
          type: array
          items:
            $ref: '#/components/schemas/ContentPage'
          description: للقصص فقط
      description: 'محتوى (قصة/فيديو/لعبة). للألعاب: إما youtube (فيديو) أو uploaded (PDF)'
    CategoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
          nullable: true
      description: ملخص فئة (حتى 3 لكل محتوى في القائمة العامة)
    AgeGroupItem:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
          example: 3-5
        ageMin:
          type: integer
        ageMax:
          type: integer
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          unique: true
        icon:
          type: string
          nullable: true
    AgeGroup:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
          unique: true
          example: 3-5
        ageMin:
          type: integer
        ageMax:
          type: integer
    AppSetting:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
          unique: true
        value:
          type: string
    Admin:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
      description: مدير (بدون كلمة المرور)
    PaginatedContent:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Content'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        page:
          type: integer
          nullable: true
    PaginatedCategories:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    PaginatedAgeGroups:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AgeGroup'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    PaginatedSettings:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AppSetting'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    PaginatedAdmins:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Admin'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    LoginBody:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: 'JWT للاستخدام في Authorization: Bearer'
        admin:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
    ContentCreateBody:
      type: object
      required:
        - title
        - type
        - ageMin
        - ageMax
        - sourceType
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        type:
          type: string
          enum:
            - story
            - video
            - game
        ageMin:
          type: integer
          minimum: 0
        ageMax:
          type: integer
          minimum: 0
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        contentUrl:
          type: string
          format: uri
          nullable: true
        fileUrl:
          type: string
          format: uri
          nullable: true
        sourceType:
          type: string
          enum:
            - uploaded
            - youtube
            - external
        isActive:
          type: boolean
        orderIndex:
          type: integer
        categoryIds:
          type: array
          items:
            type: integer
        ageGroupIds:
          type: array
          items:
            type: integer
        pages:
          type: array
          items:
            $ref: '#/components/schemas/ContentPage'
      description: 'للألعاب: sourceType=uploaded يتطلب fileUrl (PDF)، sourceType=youtube يتطلب contentUrl صالح'
...
    CategoryCreateBody:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: string
          nullable: true
...
    AgeGroupCreateBody:
      type: object
      required:
        - label
        - ageMin
        - ageMax
      properties:
        label:
          type: string
          minLength: 1
          example: 3-5
        ageMin:
          type: integer
          minimum: 0
        ageMax:
          type: integer
          minimum: 0
...
    AppSettingCreateBody:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
          minLength: 1
        value:
          type: string
...
    AdminCreateBody:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    AdminUpdateBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
paths:
  /admin/login:
    post:
      tags:
        - Admin
      summary: تسجيل دخول المدير
      description: يرجع JWT وبيانات المدير. لا يتطلب توثيقاً.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginBody'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: بريد أو كلمة مرور غير صحيحة
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/me:
    get:
      tags:
        - Admin
      summary: بيانات المدير الحالي
      security:
        - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/users:
    get:
      tags:
        - Admin
      summary: قائمة مديري النظام
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: search
          schema:
            type: string
          description: بحث في الاسم أو البريد
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAdmins'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Admin
      summary: إنشاء مدير جديد
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateBody'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: مدير بنفس البريد موجود
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/users/{id}:
    get:
      tags:
        - Admin
      summary: تفاصيل مدير بالمعرف
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Admin
      summary: تحديث مدير
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateBody'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Admin
      summary: حذف مدير
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: تم الحذف
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /age-groups:
    get:
      tags:
        - Public
      summary: قائمة الفئات العمرية (عام)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAgeGroups'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/age-groups:
    get:
      tags:
        - Admin
      summary: قائمة الفئات العمرية (إدارة)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAgeGroups'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Admin
      summary: إنشاء فئة عمرية
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgeGroupCreateBody'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgeGroup'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: فئة عمرية بنفس label موجودة
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /categories:
    get:
      tags:
        - Public
      summary: قائمة الفئات (عام)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCategories'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/categories:
    get:
      tags:
        - Admin
      summary: قائمة الفئات (إدارة)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCategories'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Admin
      summary: إنشاء فئة
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateBody'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: فئة بنفس الاسم موجودة
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /content:
    get:
      tags:
        - Public
      summary: قائمة المحتوى (عام)
      description: يرجع محتوى منشور مع فلترة عمر، نوع، فئات، بحث. كل عنصر يحتوي حتى 3 فئات.
      parameters:
        - in: query
          name: age
          schema:
            type: string
            example: 3-5
          description: نطاق عمر (مثال 3-5، 6-8)
        - in: query
          name: type
          schema:
            type: string
            enum:
              - story
              - video
              - game
          description: نوع المحتوى
        - in: query
          name: category
          schema:
            type: string
            example: 1,2,3
          description: معرفات الفئات مفصولة بفاصلة
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: بديل عن offset (يُحسب offset = (page-1)*limit)
        - in: query
          name: search
          schema:
            type: string
          description: بحث في العنوان والوصف
      responses:
        '200':
          description: قائمة منسقة مع data, total, limit, offset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedContent'
        '400':
          description: معاملات غير صالحة
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: خطأ داخلي
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/content:
    get:
      tags:
        - Admin
      summary: قائمة المحتوى (إدارة)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum:
              - story
              - video
              - game
        - in: query
          name: category
          schema:
            type: string
            example: 1,2,3
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedContent'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Admin
      summary: إنشاء محتوى
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentCreateBody'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /settings:
    get:
      tags:
        - Public
      summary: قائمة الإعدادات (عام)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSettings'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/settings:
    get:
      tags:
        - Admin
      summary: قائمة الإعدادات (إدارة)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSettings'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Admin
      summary: إنشاء إعداد
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppSettingCreateBody'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSetting'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: إعداد بنفس key موجود
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...